+++ frontend/Connect & Explore.py
@@ -104,6 +104,7 @@ def generate_dictionaries(
     return asyncio.run(generate_dictionaries_async(_cleansed_data))
 
 
+# TODO: move to utils.api.py def generate_dictionaries(cleansed_data: CleansedResult) -> DataDictionariesAndMetadata
 async def generate_dictionaries_async(
     _cleansed_data: Dict[str, Dict[str, Any]],
 ) -> Dict[str, Any]: